+++ scripts/lint/lint.sh
@@ -14,7 +14,7 @@ cd $ROOT
 FAILED=0
 
 IGNORED_FILES="uv\.lock|docs\/CARS.md|LICENSE\.md|.*\.onnx"
-IGNORED_DIRS="^third_party.*|^msgq.*|^msgq_repo.*|^opendbc.*|^opendbc_repo.*|^cereal.*|^panda.*|^rednose.*|^rednose_repo.*|^tinygrad.*|^tinygrad_repo.*|^teleoprtc.*|^teleoprtc_repo.*"
+IGNORED_DIRS="^third_party.*|^msgq.*|^msgq_repo.*|^opendbc.*|^opendbc_repo.*|^cereal.*|^panda.*|^rednose.*|^rednose_repo.*|^tinygrad.*|^tinygrad_repo.*|^teleoprtc.*|^teleoprtc_repo.*|^hoofpilot\/third_party.*"
 
 function run() {
   shopt -s extglob