+++ selfdrive/test/setup_xvfb.sh
@@ -16,4 +16,3 @@ done
 
 touch ~/.Xauthority
 export XDG_SESSION_TYPE="x11"
-xset -q